requirements.txt
blinker==1.8.2
click==8.1.7
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.1
opencv-python==4.10.0.84
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sympy==1.13.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
zipp==3.19.2
